+++ apps/governance/src/internal/governance/functionality/hooks/useProposals.ts
@@ -15,9 +15,10 @@ import {
 import { getProposals } from "../fetch";
 import { Proposal, ProposalDetailProps, ProposalProps } from "../types";
 
-const PROPOSAL_TO_REMOVE = "156";
-const removeProposal = (id: string, proposals: Proposal[]) => {
-  return proposals.filter((proposal) => proposal.id !== id);
+const PROPOSALS_TO_REMOVE = process.env.NEXT_PUBLIC_PROPOSALS_TO_REMOVE ?? "[]"
+
+const removeProposals = (proposals: Proposal[], proposalToRemove: string[]) => {
+  return proposals.filter((proposal) => !proposalToRemove.includes(proposal.id));
 };
 
 export const useProposals = (pid?: string) => {
@@ -26,17 +27,17 @@ export const useProposals = (pid?: string) => {
     queryFn: () => getProposals(),
   });
 
+  const parsedProposals = JSON.parse(PROPOSALS_TO_REMOVE) as string[]
+
   const proposals = useMemo(() => {
     const temp: ProposalProps[] = [];
     if (proposalsResponse.data !== undefined) {
-      const filtered = removeProposal(
-        PROPOSAL_TO_REMOVE,
-        proposalsResponse.data.proposals
+      const filtered = removeProposals(
+        proposalsResponse.data.proposals,
+        parsedProposals,
+        
       );
       filtered.map((item) => {
-        if (item.id === PROPOSAL_TO_REMOVE) {
-          return;
-        }
         const percents = getPercentage([
           item.final_tally_result.yes_count,
           item.final_tally_result.no_count,
@@ -64,7 +65,7 @@ export const useProposals = (pid?: string) => {
       });
     }
     return temp;
-  }, [proposalsResponse]);
+  }, [proposalsResponse, parsedProposals]);
 
   const proposalDetail = useMemo(() => {
     let temp: ProposalDetailProps = {
@@ -87,7 +88,7 @@ export const useProposals = (pid?: string) => {
     };
     if (proposalsResponse.data !== undefined) {
       const filtered = proposalsResponse.data.proposals.filter(
-        (proposal) => proposal.id === pid && proposal.id !== PROPOSAL_TO_REMOVE
+        (proposal) => proposal.id === pid && !parsedProposals.includes(proposal.id)
       );
       if (filtered.length === 0) {
         return "Proposal not found, please try again";
@@ -174,7 +175,7 @@ export const useProposals = (pid?: string) => {
       };
     }
     return temp;
-  }, [proposalsResponse, pid]);
+  }, [proposalsResponse, pid, parsedProposals]);
 
   return {
     proposals,